+++ sphinx/source/changelog.rst
@@ -15,6 +15,8 @@ Changelog
   and preserved by :ref:`skool2ctl.py`
 * :ref:`skool2bin.py` now processes :ref:`asm-if` directives (in case they
   contain :ref:`isub`, :ref:`ssub`, :ref:`ofix` or :ref:`bfix` directives)
+* Added the :ref:`remove` directive (for removing instructions at specific
+  addresses)
 * The ``LABEL`` parameter of the :ref:`label` directive may be ``*`` to provide
   the next instruction with an automatically generated label
 * Added the ``--force`` option to :ref:`skool2asm.py` (to force conversion of